rename files to match standards (no spaces, camelCase, etc)
