--- src/Components/Login Page/login.css
@@ -1,161 +1,0 @@
-.login{
-    background: #f5f5f5;
-    width: 100%;
-    height: 675px;
-    vertical-align: middle;
-}
-
-#back_image{
-    width: 100%;
-    z-index: -1;
-    height: 100%;
-    object-fit: cover;
-}
-
-#cover{
-    width: 100%;
-    height: 675px;
-    opacity: 0.5;
-    position: absolute;
-    z-index: 2;
-    color: white;
-    top: 0;
-    background: black;
-}
-
-#full-opaque-cover{
-    width: 100%;
-    height: 675px;
-    opacity: 1;
-    position: absolute;
-    z-index: 3;
-    color: white;
-    top: 0;
-    background: transparent;
-}
-
-#information{
-    width: 67.5%;
-    margin: 0 auto;
-    transition: all ease-in-out 400ms;
-    padding: 25px 0;
-    position: relative;
-}
-
-#logo{
-    margin-top: 20px;
-    font-size: 18px;
-}
-
-#twitter img{
-    width: 25px;
-    height: 25px;
-    position: absolute;
-    top: 0;
-    border-radius: 10px;
-    padding: 10px 15px;
-    cursor: pointer;
-    margin-top: 35px;
-    background: rgba(255, 255, 255, 0.0);
-    transition: all ease-in-out 400ms;
-    right: 0;
-}
-
-#twitter img:hover{
-    background: rgba(255, 255, 255, 0.2);
-    transition: all ease-in-out 400ms;
-}
-
-#big_text{
-    font-family: 'Quicksand Bold', sans-serif;
-    margin-top: 150px;
-    font-size: 50px;
-    transition: all ease-in-out 400ms;
-}
-
-#small_text{
-    font-size: 20px;
-    font-family: 'Quicksand Light', sans-serif;
-}
-
-#button{
-    margin-top: 35px;
-    font-size: 13px;
-    background: blue;
-    padding: 20px;
-    font-family: 'Quicksand', sans-serif;
-    width: 175px;
-    cursor: pointer;
-    border-radius: 2.5px;
-    text-align: center;
-    transition: all ease-in-out 400ms;
-}
-
-#button:hover{
-    box-shadow: 0 10px 10px rgba(255, 255, 255, 0.2);
-    transition: all ease-in-out 400ms;
-}
-
-#others{
-    font-size: 12px;
-    font-family: 'Quicksand Bold', sans-serif;
-    color: rgb(100, 100, 100);
-    margin: 0 auto;
-    width: 67.5%;
-    transition: all ease-in-out 400ms;
-}
-
-#others a{
-    color: rgb(100, 100, 100);
-}
-
-#others li:hover{
-    color: #9c27b0;
-    transition: all ease-in-out 400ms;
-}
-
-#others li{
-    cursor: pointer;
-    padding: 27.5px 30px 25px 0;
-    display: inline-block;
-}
-
-@media  only screen and (max-width: 767px) {
-
-    #big_text{
-        margin-top: 125px;
-        font-size: 50px;
-        transition: all ease-in-out 400ms;
-    }
-
-    #information{
-        width: 80%;
-        transition: all ease-in-out 400ms;
-    }
-
-    #others{
-        width: 80%;
-        transition: all ease-in-out 400ms;
-    }
-
-}
-
-@media  only screen and (max-width: 540px) {
-
-    #big_text{
-        margin-top: 75px;
-        font-size: 50px;
-        transition: all ease-in-out 400ms;
-    }
-
-    #information{
-        width: 80%;
-        transition: all ease-in-out 400ms;
-    }
-
-    #others{
-        width: 80%;
-        transition: all ease-in-out 400ms;
-    }
-
-}
