--- src/Components/Side Panel/sidepanel.css
@@ -1,125 +1,0 @@
-#sidepanel{
-    left: 0;
-    position: fixed;
-    margin-top: 0px;
-    margin-left: 2%;
-    width: 21%;
-    overflow: auto;
-}
-
-#active {
-    background: #f0f0f0;
-    padding: 15px;
-    position: relative;
-    margin: 10px 0;
-    transition: all ease-in-out 400ms;
-    border-radius: 5px;
-}
-
-a{
-    text-decoration: none;
-    color: black;
-}
-
-#link{
-    padding: 15px;
-    position: relative;
-    margin: 10px 0;
-    transition: all ease-in-out 400ms;
-    border-radius: 5px;
-}
-
-#report_navigation_title{
-    cursor: pointer;
-}
-
-#expand{
-    position: absolute;
-    right: 0;
-    margin-right: 25px;
-    cursor: pointer;
-}
-
-#mobile-expand-icon{
-    display: none;
-}
-
-#mini-links{
-    font-family: 'Quicksand Light', sans-serif;
-    display: block;
-    transition: all ease-in-out 400ms;
-}
-
-#mini-links li{
-    list-style-type: none;
-    line-height: 1.8em;
-    cursor: pointer;
-    margin-left: -25px;
-}
-
-#no_links_display{
-    display: none;
-    transition: all ease-in-out 400ms;
-}
-
-#sidepanel-mobile{
-    display: none;
-}
-
-@media  only screen and (max-width: 767px) {
-    #sidepanel-mobile{
-        left: 0;
-        position: fixed;
-        margin-top: 0px;
-        margin-left: 0;
-        width: 100%;
-        display: none;
-        top: 0;
-        height: 100%;
-        padding-top: 150px;
-        z-index: 99999999999999999999;
-        background: rgba(255, 255, 255, 0.9);
-        overflow: auto;
-    }
-
-    #sidepanel{
-        display: none;
-    }
-
-    #active {
-        margin: 25px auto;
-        width: 50%;
-    }
-    
-    a{
-        text-decoration: none;
-        color: black;
-    }
-    
-    #link{
-        margin: 25px auto;
-        width: 50%;
-    }
-
-    #mobile-expand-icon{
-        display: block;
-        border: solid #fafafa 1px;
-        box-shadow: 0 0 10px rgba(0, 0, 0, 0.0);
-        padding: 8.5px 17.5px;
-        border-radius: 0px;
-        position: fixed;
-        background: transparent;
-        top: 0;
-        z-index: 9999999999999999999999;
-        left: 0;
-        cursor: pointer;
-        font-size: 20px;
-        margin-top: 15px;
-        margin-left: 5px;
-    }
-
-    #menu_bar {
-        width: 17.5px;
-        height: 17.5px;
-    }
-}
